+++ examples/trait-incrementer/lib.rs
@@ -3,29 +3,9 @@
 
 use ink_lang as ink;
 
-/// Allows to increment and get the current value.
-#[ink::trait_definition]
-pub trait Increment {
-    /// Increments the current value of the implementer by one (1).
-    #[ink(message)]
-    fn inc(&mut self);
-
-    /// Returns the current value of the implementer.
-    #[ink(message)]
-    fn get(&self) -> u64;
-}
-
-/// Allows to reset the current value.
-#[ink::trait_definition]
-pub trait Reset {
-    /// Resets the current value to zero.
-    #[ink(message)]
-    fn reset(&mut self);
-}
-
 #[ink::contract]
 pub mod incrementer {
-    use super::{
+    use traits::{
         Increment,
         Reset,
     };
